feat: 그리드 구독 스낵바 기능 utils로 빼기
